refactor: 기존 ApplicationPagination을 ApplicationList로 컴포넌트 네이밍 수정
